+++ test/i18n/steps/es.ts
@@ -1,0 +1,15 @@
+import { createBdd } from 'playwright-bdd';
+
+const { Given, When, Then } = createBdd();
+
+Given('que abro la url {string}', async () => {
+  // noop
+});
+
+When('hago clic en el enlace {string}', async function () {
+  // noop
+});
+
+Then('veo en el título {string}', async function () {
+  // noop
+});