+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Dev
+* feature: provide full localized step titles to Playwright HTML reporter ([#229](https://github.com/vitalets/playwright-bdd/issues/229))
+
 ## 7.5.0
 * support external attachments in Cucumber HTML reporter ([#182](https://github.com/vitalets/playwright-bdd/issues/182))
 * support trace-viewer in Cucumber HTML reporter ([#182](https://github.com/vitalets/playwright-bdd/issues/182))